refactored category controller and service

+++ src/main/java/com/parovsky/traver/service/impl/CategoryServiceImpl.java
@@ -1,10 +1,13 @@
 package com.parovsky.traver.service.impl;
 
 import com.parovsky.traver.dao.CategoryDAO;
+import com.parovsky.traver.dao.UserDAO;
 import com.parovsky.traver.dto.CategoryDTO;
 import com.parovsky.traver.entity.Category;
+import com.parovsky.traver.exception.impl.CategoryIsAlreadyExistException;
+import com.parovsky.traver.exception.impl.CategoryNotFoundException;
 import com.parovsky.traver.service.CategoryService;
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.AllArgsConstructor;
 import org.springframework.lang.NonNull;
 import org.springframework.stereotype.Service;
 
@@ -12,14 +15,12 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 @Service
+@AllArgsConstructor(onConstructor = @__({@org.springframework.beans.factory.annotation.Autowired}))
 public class CategoryServiceImpl implements CategoryService {
 
 	private final CategoryDAO categoryDAO;
 
-	@Autowired
-	public CategoryServiceImpl(CategoryDAO categoryDAO) {
-		this.categoryDAO = categoryDAO;
-	}
+	private final UserDAO userDAO;
 
 	@Override
 	public List<CategoryDTO> getAllCategories() {
@@ -28,39 +29,42 @@ public class CategoryServiceImpl implements CategoryService {
 	}
 
 	@Override
-	public List<CategoryDTO> getFavoriteCategories(Long userId) {
-		List<Category> categories = categoryDAO.getFavoriteCategories(userId);
+	public List<CategoryDTO> getFavoriteCategories() {
+		String userEmail = userDAO.getCurrentUserEmail();
+		List<Category> categories = categoryDAO.getFavoriteCategories(userEmail);
 		return categories.stream().map(CategoryService::transformCategoryToCategoryDTO).collect(Collectors.toList());
 	}
 
 	@Override
-	public CategoryDTO getCategoryById(Long id) {
+	public CategoryDTO getCategoryById(@NonNull Long id) throws CategoryNotFoundException {
 		Category category = categoryDAO.getCategoryById(id);
+		if (category == null) {
+			throw new CategoryNotFoundException();
+		}
 		return CategoryService.transformCategoryToCategoryDTO(category);
 	}
 
 	@Override
-	public boolean isCategoryExistById(Long id) {
-		return categoryDAO.isCategoryExistById(id);
-	}
-
-	@Override
-	public boolean isCategoryExistByName(String name) {
-		return categoryDAO.isCategoryExistByName(name);
-	}
-
-	@Override
-	public CategoryDTO saveCategory(@NonNull CategoryDTO categoryDTO) {
+	public CategoryDTO saveCategory(@NonNull CategoryDTO categoryDTO) throws CategoryIsAlreadyExistException {
+		if (categoryDAO.isCategoryExistByName(categoryDTO.getName())) {
+			throw new CategoryIsAlreadyExistException();
+		}
 		return CategoryService.transformCategoryToCategoryDTO(categoryDAO.saveCategory(categoryDTO));
 	}
 
 	@Override
-	public CategoryDTO updateCategory(@NonNull CategoryDTO categoryDTO) {
+	public CategoryDTO updateCategory(@NonNull CategoryDTO categoryDTO) throws CategoryNotFoundException {
+		if (!categoryDAO.isCategoryExistById(categoryDTO.getId())) {
+			throw new CategoryNotFoundException();
+		}
 		return CategoryService.transformCategoryToCategoryDTO(categoryDAO.updateCategory(categoryDTO));
 	}
 
 	@Override
-	public void deleteCategory(Long id) {
+	public void deleteCategory(@NonNull Long id) throws CategoryNotFoundException {
+		if (!categoryDAO.isCategoryExistById(id)) {
+			throw new CategoryNotFoundException();
+		}
 		categoryDAO.deleteCategory(id);
 	}
 }
